retry a few times if the file number of html report is suspicious
should close #14.

it appears that the OLMS web server sometime returns the wrong response
when we ask for the detailed html form. so we check that the file number
is the same as the srNum (which comes from the base info), and if it
is not we retry three times.

for the cases that it gave us the wrong response, this seems to work
the very second retry.

unfortunately, there are some cases where the srNum and the file number
are validly different, and so after three retries, we accept the info
from the html response.